Rework SP key and certificate, and integrate with metadata
The existing metadata, key and certificate was from other test apps. So
it was time to generate some new and then tweak the metadata to work
with my test Shibboleth setup. The main thing being to include the
matching certs/pub keys in the SP meta data so that I don't need to
also install on IDP.

I also renamed the IDP and SP metadata file just to be tidy and
abundantly clear.